Warning corrections
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@35317 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ include/wx/univ/notebook.h
@@ -68,7 +68,7 @@ public:
     // --------------------------------------
 
     virtual int SetSelection(size_t nPage);
-    virtual int GetSelection() const { return m_sel; }
+    virtual int GetSelection() const { return (int) m_sel; }
 
     virtual bool SetPageText(size_t nPage, const wxString& strText);
     virtual wxString GetPageText(size_t nPage) const;
